move contstants file
