+++ config/constants.js
@@ -1,0 +1,9 @@
+const supportedThemes = [
+    'sap_fiori_3',
+    'sap_fiori_3_dark',
+    'sap_fiori_3_hcb',
+    'sap_fiori_3_hcw',
+    'sap_fiori_3_light_dark'
+];
+
+module.exports = supportedThemes;
